Update group button styles
