+++ src/components/Chart.tsx
@@ -1,37 +1,9 @@
 import { LineChart } from "@mui/x-charts/LineChart";
 import { useContext } from "react";
 import { stockContext } from "../App";
-import getSixWeeksStockData from "../helpers/getSixWeeksStockData";
-import getFiveMonthsStockData from "../helpers/getFiveMonthsStockData";
-import getSevenDaysStockData from "../helpers/getSevenDaysStockData";
-import getStoredDataFromStorage from "../helpers/getStoredDataFromStorage";
-import Button from "./Button/Button";
-import ButtonGroup from "./ButtonGroup/ButtonGroup";
 
 export default function Chart() {
-  const stocksData = getStoredDataFromStorage();
-
-  const { stockHistoricalData, updateStockHistoricalData } =
-    useContext(stockContext);
-
-  function updateToSevenDays() {
-    const sevenDaysOfStock = getSevenDaysStockData(stocksData!);
-    if (sevenDaysOfStock) {
-      updateStockHistoricalData(sevenDaysOfStock);
-    }
-  }
-  function updateToSixWeeks() {
-    const sixWeeksOfStocks = getSixWeeksStockData(stocksData!);
-    if (sixWeeksOfStocks) {
-      updateStockHistoricalData(sixWeeksOfStocks);
-    }
-  }
-  function updateToFiveMonths() {
-    const fiveMonthsOfStocks = getFiveMonthsStockData(stocksData!);
-    if (fiveMonthsOfStocks) {
-      updateStockHistoricalData(fiveMonthsOfStocks);
-    }
-  }
+  const { stockHistoricalData } = useContext(stockContext);
 
   function renderChart() {
     if (!stockHistoricalData) {
@@ -78,14 +50,5 @@ export default function Chart() {
     );
   }
 
-  return (
-    <>
-      <ButtonGroup>
-        <Button onClick={updateToSevenDays}>7 days</Button>
-        <Button onClick={updateToSixWeeks}>6 Weeks</Button>
-        <Button onClick={updateToFiveMonths}>5 months</Button>
-      </ButtonGroup>
-      {renderChart()}
-    </>
-  );
+  return <>{renderChart()}</>;
 }
